// src/lib/api-bal/generated/types.gen.ts
export type BaseLocaleSync = {
    status: 'outdated' | 'synced' | 'conflict';
    isPaused: boolean;
    lastUploadedRevisionId: string;
    currentUpdated: string;
};

export type BaseLocaleFondDeCarte = {
    name: string;
    url: string;
};

export type BaseLocaleSetting = {
    otherBalPublishedIgnored: boolean;
    languageGoalIgnored: boolean;
    toponymeGoalIgnored: boolean;
    fondsDeCartes: Array<BaseLocaleFondDeCarte>;
    ignoredAlertCodes: Array<string>;
};

export type Numero = {
    id: string;
    banId: string;
    createdAt: string;
    updatedAt: string;
    deletedAt: string;
    balId: string;
    voieId: string;
    toponymeId: string;
    numero: number;
    suffixe: string;
    numeroComplet: string;
    comment: string;
    parcelles: Array<string>;
    certifie: boolean;
    communeDeleguee: string;
    positions: Array<Position>;
    baseLocale: BaseLocale;
    voie: Voie;
    toponyme: Toponyme;
};

export type Voie = {
    id: string;
    banId: string;
    createdAt: string;
    updatedAt: string;
    deletedAt: string;
    balId: string;
    nom: string;
    nomAlt: {
        [key: string]: unknown;
    };
    typeNumerotation: 'numerique' | 'metrique';
    centroid: {
        [key: string]: unknown;
    };
    trace: {
        [key: string]: unknown;
    };
    bbox: Array<number>;
    codeVoie: string;
    comment: string;
    baseLocale: BaseLocale;
    numeros: Array<Numero>;
};

export type BaseLocale = {
    id: string;
    banId: string;
    createdAt: string;
    updatedAt: string;
    deletedAt: string;
    nom: string;
    communeNom?: string;
    communeNomsAlt: {
        [key: string]: unknown;
    };
    commune: string;
    emails: Array<string>;
    token: string;
    status: 'draft' | 'published' | 'demo' | 'replaced';
    importType: 'api-depot' | 'ban' | 'csv';
    habilitationId: string;
    sync: BaseLocaleSync;
    settings: BaseLocaleSetting;
    voies: Array<Voie>;
    toponymes: Array<Toponyme>;
    numeros: Array<Numero>;
};

export type Toponyme = {
    id: string;
    banId: string;
    createdAt: string;
    updatedAt: string;
    deletedAt: string;
    balId: string;
    nom: string;
    nomAlt: {
        [key: string]: unknown;
    };
    communeDeleguee: string;
    parcelles: Array<string>;
    positions: Array<Position>;
    codeVoie: string;
    baseLocale: BaseLocale;
    numeros: Array<Numero>;
};

export type Position = {
    id: string;
    toponymeId: string;
    numeroId: string;
    type: 'entrée' | 'bâtiment' | 'cage d’escalier' | 'logement' | 'service technique' | 'délivrance postale' | 'parcelle' | 'segment';
    source: string;
    rank: number;
    point: {
        [key: string]: unknown;
    };
    toponyme: Toponyme;
    numero: Numero;
};